refactor: move tab-stepper from molecules to templates
